third_party/blink/web_tests/platform/generic/fast/speechsynthesis/speech-synthesis-utterance-uses-voice-expected.txt

This tests that we can get synthesizer voices on the Mac

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Verify that setting and retrieving the voice works.
PASS u.voice.name == voice.name is true
Speech job finished successfully.
PASS successfullyParsed is true

TEST COMPLETE
